fix(comp): resolution of requirements
