+++ src/components/Input/styles.scss
@@ -64,10 +64,7 @@ $box-shadow-pink: rgba(246, 43, 253, 0.65);
   &Face {
     position: absolute;
     left: 0;
-
-    &Blue {
-      background: $linear-gradient-blue;
-    }
+    background: $linear-gradient-blue;
 
     &Pink {
       background: $linear-gradient-pink;
@@ -82,11 +79,8 @@ $box-shadow-pink: rgba(246, 43, 253, 0.65);
     transform-origin: 50% 0;
     border-radius: 2px;
     box-sizing: border-box;
-
-    &Blue {
-      border: 1px solid $color-border-botton-blue;
-      box-shadow: 0px 0px 6px 2px $box-shadow-blue;
-    }
+    border: 1px solid $color-border-botton-blue;
+    box-shadow: 0px 0px 6px 2px $box-shadow-blue;
 
     &Pink {
       border: 1px solid $color-border-botton-pink;
